3_2_redes2/practica2/requirements.txt
argparse
pycryptodomex
requests
